Separate logical location header
